changed previous fix so that now you can optional skip registering the UI routes for swagger.
This allows you to just put the swagger ui in your asp.net mvc project.

The reason for this was that for best performance you don't want to set runAllManagedModulesForAllRequests=true but swashbuckle
was forcing this so that the embedded static files could be delivered. this was causing small performance issues to the rest of the site which
really didn't need static files going through the asp.net pipeline