Removes default ports from `DefaultRootUrlResolver` when applicable
I've noticed that regardless of where the swagger documents are hosted,
the URI of the swagger spec always includes the port, even if it uses
the protocol default port i.e. 80 for http and 443 for https.

This update:
* alters the `DefaultRootUrlResolver` to use a `UriBuilder` handle the
decision of including the port of not (instead of using string
concatenation)
* adds new tests for omitting the port from the url when applicable

This will result in urls like `http://tempuri.org/swaggger/docs/1`
instead of `http://tempuri.org:80/swaggger/docs/1`.